Move bootstrap_user to ilsgateway
