+++ custom/ilsgateway/tests/handlers/utils.py
@@ -13,7 +13,7 @@ from corehq.toggles import USE_SMS_WITH_INACTIVE_CONTACTS
 from custom.ilsgateway.models import ILSGatewayConfig
 from custom.ilsgateway.utils import make_loc
 from custom.logistics.tests.test_script import TestScript
-from custom.logistics.tests.utils import bootstrap_user
+from custom.ilsgateway.tests.utils import bootstrap_user
 from casexml.apps.stock.models import DocDomainMapping
 from six.moves import range
